backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
bcrypt==4.1.3
black==25.9.0
boto3==1.40.55
botocore==1.40.55
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cors==1.0.1
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.109.1
flake8==7.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
isort==7.0.0
jmespath==1.0.1
# jq==1.10.0  # Nécessite compilation C++, désactivé temporairement
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
postgrest==2.22.1
propcache==0.4.1
pyasn1==0.4.8
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.6.3
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.4.0
python-multipart==0.0.18
pytokens==0.2.0
pytz==2023.3
realtime==2.22.1
reportlab==4.2.5
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.35.0
storage3==2.22.1
StrEnum==0.4.15
supabase==2.22.1
supabase-auth==2.22.1
supabase-functions==2.22.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0

# Production Server  
gunicorn==23.0.0
# gevent==24.11.1  # Nécessite compilation C++, désactivé temporairement
# aiohttp==3.12.14  # Nécessite compilation C++, remplacé par httpx
aiohttp-cors==0.7.0

# Production-Grade Monitoring & Security
sentry-sdk==1.40.0
structlog==23.3.0
psutil==5.9.8
redis==5.0.1
stripe==11.2.0

# 2FA & Security
pyotp==2.9.0
qrcode==7.4.2
# Pillow==11.0.0  # Python 3.14 pas encore supporté, utiliser version binaire
pypng==0.20220715.0  # Alternative légère pour QR codes

# Email & Templates
Jinja2==3.1.3
celery==5.3.6
sendgrid==6.11.0
openai==1.12.0

# AI
openai==1.12.0
scikit-learn>=1.3.2

# Notifications
twilio==8.10.0
slack_sdk==3.26.1
firebase-admin==6.4.0
